# BufferedStream performance improvements