The following changes improve `BufferedStream` perf a bit and in the near future might allow for the implementation of `BufferedFileStreamStrategy`, where `BufferedStream` is going to be used for all the buffering logic on top of actual `FileStreamStrategy` (which derives from `Stream`).

Changes:

1. Move throws to local functions to allow for the inlining of small guard methods.
2. ~~Use `_buffer.Length` instead of `_bufferLength` as `BufferedStream` might resize `_buffer` when it decides it's a good idea ([source code](https://github.com/dotnet/runtime/blob/d5ab93c4a8e45da2dab8924c2165000bf78f84e6/src/libraries/System.Private.CoreLib/src/System/IO/BufferedStream.cs#L115)).~~
3. ~~Don't try to acquire the lock if it already has been acquired~~
4. Add common code path for `ReadAsync` and `WriteAsync` when there is nothing to Flush and buffering is not beneficial (user buffer size => internal buffer size). 
5. Optimize common `WriteByte` code path (at least one byte has already been written to buffer and we don't need to perform all expensive checks). [ReadByte](https://github.com/dotnet/runtime/blob/d5ab93c4a8e45da2dab8924c2165000bf78f84e6/src/libraries/System.Private.CoreLib/src/System/IO/BufferedStream.cs#L755-L760) already has such an optimization