@adamsitnik 2021-02-24
> Can you share perf numbers? 

Sure! For the following benchmarks where `BufferedStream` just wraps a `FileStream` with buffering disabled (bufferSize set to 1):

<details>

// Licensed to the .NET Foundation under one or more agreements.
// The .NET Foundation licenses this file to you under the MIT license.
// See the LICENSE file in the project root for more information.

using System.Collections.Generic;
using System.Linq;
using System.Threading;
using System.Threading.Tasks;
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Extensions;
using MicroBenchmarks;

namespace System.IO
{
    [BenchmarkCategory(Categories.Libraries)]
    public class BufferedStreamTests
    {
        private const int OneKibibyte = 1 << 10; // 1024
        private const int HalfKibibyte = OneKibibyte >> 1;
        private const int FourKibibytes = OneKibibyte << 2; // default Stream buffer size
        private const int OneMibibyte = OneKibibyte << 10;
        private const int HundredMibibytes = OneMibibyte * 100;

        private Dictionary<long, string> _sourceFilePaths, _destinationFilePaths;

        private Dictionary<int, byte[]> _userBuffers;

        private void Setup(params long[] fileSizes)
        {
            _userBuffers = new Dictionary<int, byte[]>()
            {
                { HalfKibibyte, ValuesGenerator.Array<byte>(HalfKibibyte) },
                { FourKibibytes, ValuesGenerator.Array<byte>(FourKibibytes) },
            };
            _sourceFilePaths = fileSizes.ToDictionary(size => size, size => CreateFileWithRandomContent(size));
            _destinationFilePaths = fileSizes.ToDictionary(size => size, size => CreateFileWithRandomContent(size));

            static string CreateFileWithRandomContent(long fileSize)
            {
                string filePath = FileUtils.GetTestFilePath();
                File.WriteAllBytes(filePath, ValuesGenerator.Array<byte>((int)fileSize));
                return filePath;
            }
        }

        [GlobalCleanup]
        public void Cleanup()
        {
            foreach (string filePath in _sourceFilePaths.Values.Concat(_destinationFilePaths.Values))
            {
                File.Delete(filePath);
            }
        }

        [GlobalSetup(Targets = new[] { nameof(ReadByte), nameof(WriteByte) })]
        public void SetuOneKibibyteBenchmarks() => Setup(OneKibibyte);

        [Benchmark]
        [Arguments(OneKibibyte, FileOptions.None)]
        [Arguments(OneKibibyte, FileOptions.Asynchronous)] // calling ReadByte() on bigger files makes no sense, so we don't have more test cases
        public int ReadByte(long fileSize, FileOptions options)
        {
            int result = default;
            using (FileStream fileStream = new FileStream(_sourceFilePaths[fileSize], FileMode.Open, FileAccess.Read, FileShare.Read, 1, options))
            using (BufferedStream bufferedStream = new BufferedStream(fileStream, FourKibibytes))
            {
                for (long i = 0; i < fileSize; i++)
                {
                    result += bufferedStream.ReadByte();
                }
            }

            return result;
        }

        [Benchmark]
        [Arguments(OneKibibyte, FileOptions.None)]
        [Arguments(OneKibibyte, FileOptions.Asynchronous)]
        public void WriteByte(long fileSize, FileOptions options)
        {
            using (FileStream fileStream = new FileStream(_destinationFilePaths[fileSize], FileMode.Create, FileAccess.Write, FileShare.Read, 1, options))
            using (BufferedStream bufferedStream = new BufferedStream(fileStream, FourKibibytes))
            {
                for (int i = 0; i < fileSize; i++)
                {
                    bufferedStream.WriteByte(default);
                }
            }
        }

        [GlobalSetup(Targets = new[] { nameof(Read), "ReadAsync", nameof(Write), "WriteAsync" })]
        public void SetupBigFileBenchmarks() => Setup(OneKibibyte, OneMibibyte, HundredMibibytes);

        [Benchmark]
        [Arguments(OneKibibyte, HalfKibibyte, FileOptions.None)] // userBufferSize is less than StreamBufferSize, buffering makes sense
        [Arguments(OneKibibyte, FourKibibytes, FileOptions.None)] // the buffer provided by User and internal Stream buffer are of the same size, buffering makes NO sense
        [Arguments(OneMibibyte, HalfKibibyte, FileOptions.None)]
        [Arguments(OneMibibyte, FourKibibytes, FileOptions.None)]
        [Arguments(HundredMibibytes, HalfKibibyte, FileOptions.None)]
        [Arguments(HundredMibibytes, FourKibibytes, FileOptions.None)]
        public long Read(long fileSize, int userBufferSize, FileOptions options)
        {
            byte[] userBuffer = _userBuffers[userBufferSize];
            long bytesRead = 0;
            using (FileStream fileStream = new FileStream(_sourceFilePaths[fileSize], FileMode.Open, FileAccess.Read, FileShare.Read, 1, options))
            using (BufferedStream bufferedStream = new BufferedStream(fileStream, FourKibibytes))
            {
                while (bytesRead < fileSize)
                {
                    bytesRead += bufferedStream.Read(userBuffer, 0, userBuffer.Length);
                }
            }

            return bytesRead;
        }

        [Benchmark]
        [Arguments(OneKibibyte, HalfKibibyte, FileOptions.None)]
        [Arguments(OneKibibyte, FourKibibytes, FileOptions.None)]
        [Arguments(OneMibibyte, HalfKibibyte, FileOptions.None)]
        [Arguments(OneMibibyte, FourKibibytes, FileOptions.None)]
        [Arguments(HundredMibibytes, HalfKibibyte, FileOptions.None)]
        [Arguments(HundredMibibytes, FourKibibytes, FileOptions.None)]
        public void Write(long fileSize, int userBufferSize, FileOptions options)
        {
            byte[] userBuffer = _userBuffers[userBufferSize];
            using (FileStream fileStream = new FileStream(_destinationFilePaths[fileSize], FileMode.Create, FileAccess.Write, FileShare.Read, 1, options))
            using (BufferedStream bufferedStream = new BufferedStream(fileStream, FourKibibytes))
            {
                for (int i = 0; i < fileSize / userBufferSize; i++)
                {
                    bufferedStream.Write(userBuffer, 0, userBuffer.Length);
                }
            }
        }

#if !NETFRAMEWORK // APIs added in .NET Core 2.0
        [Benchmark]
        [Arguments(OneKibibyte, HalfKibibyte, FileOptions.Asynchronous)]
        [Arguments(OneKibibyte, HalfKibibyte, FileOptions.None)] // common use case (sync open, later async usage)
        [Arguments(OneKibibyte, FourKibibytes, FileOptions.Asynchronous)]
        [Arguments(OneKibibyte, FourKibibytes, FileOptions.None)]
        [Arguments(OneMibibyte, HalfKibibyte, FileOptions.Asynchronous)]
        [Arguments(OneMibibyte, HalfKibibyte, FileOptions.None)]
        [Arguments(OneMibibyte, FourKibibytes, FileOptions.Asynchronous)]
        [Arguments(OneMibibyte, FourKibibytes, FileOptions.None)]
        [Arguments(HundredMibibytes, HalfKibibyte, FileOptions.Asynchronous)]
        [Arguments(HundredMibibytes, HalfKibibyte, FileOptions.None)]
        [Arguments(HundredMibibytes, FourKibibytes, FileOptions.Asynchronous)]
        [Arguments(HundredMibibytes, FourKibibytes, FileOptions.None)]
        [BenchmarkCategory(Categories.NoWASM)]
        public async Task<long> ReadAsync(long fileSize, int userBufferSize, FileOptions options)
        {
            CancellationToken cancellationToken = CancellationToken.None;
            Memory<byte> userBuffer = new Memory<byte>(_userBuffers[userBufferSize]);
            long bytesRead = 0;
            using (FileStream fileStream = new FileStream(_sourceFilePaths[fileSize], FileMode.Open, FileAccess.Read, FileShare.Read, 1, options))
            using (BufferedStream bufferedStream = new BufferedStream(fileStream, FourKibibytes))
            {
                while (bytesRead < fileSize)
                {
                    bytesRead += await bufferedStream.ReadAsync(userBuffer, cancellationToken);
                }
            }

            return bytesRead;
        }

        [Benchmark]
        [Arguments(OneKibibyte, HalfKibibyte, FileOptions.Asynchronous)]
        [Arguments(OneKibibyte, HalfKibibyte, FileOptions.None)]
        [Arguments(OneKibibyte, FourKibibytes, FileOptions.Asynchronous)]
        [Arguments(OneKibibyte, FourKibibytes, FileOptions.None)]
        [Arguments(OneMibibyte, HalfKibibyte, FileOptions.Asynchronous)]
        [Arguments(OneMibibyte, HalfKibibyte, FileOptions.None)]
        [Arguments(OneMibibyte, FourKibibytes, FileOptions.Asynchronous)]
        [Arguments(OneMibibyte, FourKibibytes, FileOptions.None)]
        [Arguments(HundredMibibytes, HalfKibibyte, FileOptions.Asynchronous)]
        [Arguments(HundredMibibytes, HalfKibibyte, FileOptions.None)]
        [Arguments(HundredMibibytes, FourKibibytes, FileOptions.Asynchronous)]
        [Arguments(HundredMibibytes, FourKibibytes, FileOptions.None)]
        [BenchmarkCategory(Categories.NoWASM)]
        public async Task WriteAsync(long fileSize, int userBufferSize, FileOptions options)
        {
            CancellationToken cancellationToken = CancellationToken.None;
            Memory<byte> userBuffer = new Memory<byte>(_userBuffers[userBufferSize]);
            using (FileStream fileStream = new FileStream(_destinationFilePaths[fileSize], FileMode.Create, FileAccess.Write, FileShare.Read, 1, options))
            using (BufferedStream bufferedStream = new BufferedStream(fileStream, FourKibibytes))
            {
                for (int i = 0; i < fileSize / userBufferSize; i++)
                {
                    await bufferedStream.WriteAsync(userBuffer, cancellationToken);
                }
            }
        }
#endif
    }
}

</details>

I got the following improvements:

WriteByte: +2-4%

|    Method |           Toolchain | fileSize |      options |       Mean | Ratio |
|---------- |-------------------- |--------- |------------- |-----------:|------:|
| WriteByte |  \after\CoreRun.exe |     1024 |         None |   993.7 us |  0.98 |
| WriteByte | \before\CoreRun.exe |     1024 |         None | 1,011.4 us |  1.00 |
|           |                     |          |              |            |       |
| WriteByte |  \after\CoreRun.exe |     1024 | Asynchronous | 1,066.9 us |  0.96 |
| WriteByte | \before\CoreRun.exe |     1024 | Asynchronous | 1,106.8 us |  1.00 |

Read: +0-5%

|     Method |           Toolchain |  fileSize | userBufferSize |      options |            Mean | Ratio |
|----------- |-------------------- |---------- |--------------- |------------- |----------------:|------:|
|       Read |  \after\CoreRun.exe |   1048576 |           4096 |         None |       658.45 us |  0.98 |
|       Read | \before\CoreRun.exe |   1048576 |           4096 |         None |       674.29 us |  1.00 |
|            |                     |           |                |              |                 |       |
|      Write |  \after\CoreRun.exe |   1048576 |           4096 |         None |     5,705.51 us |  0.97 |
|      Write | \before\CoreRun.exe |   1048576 |           4096 |         None |     5,875.82 us |  1.00 |
|            |                     |           |                |              |                 |       |
|       Read |  \after\CoreRun.exe | 104857600 |            512 |         None |    86,316.24 us |  0.95 |
|       Read | \before\CoreRun.exe | 104857600 |            512 |         None |    90,941.06 us |  1.00 |
|            |                     |           |                |              |                 |       |
|      Write |  \after\CoreRun.exe | 104857600 |            512 |         None |   232,243.21 us |  0.98 |
|      Write | \before\CoreRun.exe | 104857600 |            512 |         None |   234,376.06 us |  1.00 |

WriteAsync: +0-16% (see the allocated column)

|     Method |           Toolchain |  fileSize | userBufferSize |      options |            Mean | Ratio |    Allocated |
|----------- |-------------------- |---------- |--------------- |------------- |----------------:|------:|-------------:|
| WriteAsync |  \after\CoreRun.exe |   1048576 |           4096 |         None |     6,627.60 us |  0.94 |     29,333 B |
| WriteAsync | \before\CoreRun.exe |   1048576 |           4096 |         None |     7,091.43 us |  1.00 |     74,385 B |
|            |                     |           |                |              |                 |       |              |
| WriteAsync |  \after\CoreRun.exe |   1048576 |           4096 | Asynchronous |    16,005.07 us |  0.98 |     80,482 B |
| WriteAsync | \before\CoreRun.exe |   1048576 |           4096 | Asynchronous |    16,375.12 us |  1.00 |    125,533 B |
|            |                     |           |                |              |                 |       |              |
| WriteAsync |  \after\CoreRun.exe | 104857600 |           4096 |         None |   294,321.83 us |  0.84 |  2,869,760 B |
| WriteAsync | \before\CoreRun.exe | 104857600 |           4096 |         None |   351,712.96 us |  1.00 |  7,375,096 B |
|            |                     |           |                |              |                 |       |              |
| WriteAsync |  \after\CoreRun.exe | 104857600 |           4096 | Asynchronous | 1,560,653.26 us |  1.00 |  7,988,960 B |
| WriteAsync | \before\CoreRun.exe | 104857600 |           4096 | Asynchronous | 1,564,277.74 us |  1.00 | 12,493,112 B |

ReadAsync: +0-28%

|     Method |           Toolchain |  fileSize | userBufferSize |      options |            Mean | Ratio |    Allocated |
|----------- |-------------------- |---------- |--------------- |------------- |----------------:|------:|-------------:|
|  ReadAsync |  \after\CoreRun.exe |      1024 |           4096 |         None |        64.67 us |  0.98 |        776 B |
|  ReadAsync | \before\CoreRun.exe |      1024 |           4096 |         None |        66.18 us |  1.00 |        984 B |
|            |                     |           |                |              |                 |       |              |
|  ReadAsync |  \after\CoreRun.exe |      1024 |           4096 | Asynchronous |        83.97 us |  0.98 |        920 B |
|  ReadAsync | \before\CoreRun.exe |      1024 |           4096 | Asynchronous |        85.72 us |  1.00 |      1,128 B |
|            |                     |           |                |              |                 |       |              |
|  ReadAsync |  \after\CoreRun.exe |   1048576 |            512 | Asynchronous |     5,051.38 us |  0.94 |    137,851 B |
|  ReadAsync | \before\CoreRun.exe |   1048576 |            512 | Asynchronous |     5,350.31 us |  1.00 |    137,849 B |
|            |                     |           |                |              |                 |       |              |
|  ReadAsync |  \after\CoreRun.exe |   1048576 |           4096 |         None |     1,077.96 us |  0.72 |     29,337 B |
|  ReadAsync | \before\CoreRun.exe |   1048576 |           4096 |         None |     1,498.27 us |  1.00 |     82,584 B |
|            |                     |           |                |              |                 |       |              |
|  ReadAsync |  \after\CoreRun.exe |   1048576 |           4096 | Asynchronous |     4,707.53 us |  0.90 |     80,482 B |
|  ReadAsync | \before\CoreRun.exe |   1048576 |           4096 | Asynchronous |     5,228.57 us |  1.00 |    133,731 B |
|            |                     |           |                |              |                 |       |              |
|  ReadAsync |  \after\CoreRun.exe | 104857600 |           4096 |         None |   164,126.66 us |  0.84 |  2,868,008 B |
|  ReadAsync | \before\CoreRun.exe | 104857600 |           4096 |         None |   196,426.32 us |  1.00 |  8,192,696 B |
|            |                     |           |                |              |                 |       |              |
|  ReadAsync |  \after\CoreRun.exe | 104857600 |           4096 | Asynchronous |   492,767.83 us |  0.97 |  7,988,688 B |
|  ReadAsync | \before\CoreRun.exe | 104857600 |           4096 | Asynchronous |   509,748.57 us |  1.00 | 13,313,008 B |